+++ Firmware/CHANGELOG.md
@@ -2,10 +2,12 @@
 Please add a note of your changes below this heading if you make a PR
 
 ### Added
+* **Storing of configuration parameters to Non Volatile Memory**
 * Reporting error if your encoder CPR is incorrect
 * Reporting of DRV status/control registers and fault codes
 * DRV status read script
 * Microsecond delay function
+* Ability to start anticogging calibration over USB protocol
 * Travis-CI
 
 ### Changed